+++ starlite_users/main.py
@@ -118,7 +118,11 @@ class StarliteUsers:
                     logout_path=self._config.auth_handler_config.logout_path,
                     user_read_dto=self._config.user_read_dto,
                     service_dependency=get_service_dependency(
-                        self._config.user_model, self._config.role_model, self._config.user_service_class
+                        user_model=self._config.user_model,
+                        role_model=self._config.role_model,
+                        user_service_class=self._config.user_service_class,
+                        secret=self._config.secret,
+                        hash_schemes=self._config.hash_schemes,
                     ),
                     auth_backend=auth_backend,
                 )
@@ -130,7 +134,11 @@ class StarliteUsers:
                     user_read_dto=self._config.user_read_dto,
                     user_update_dto=self._config.user_update_dto,
                     service_dependency=get_service_dependency(
-                        self._config.user_model, self._config.role_model, self._config.user_service_class
+                        user_model=self._config.user_model,
+                        role_model=self._config.role_model,
+                        user_service_class=self._config.user_service_class,
+                        secret=self._config.secret,
+                        hash_schemes=self._config.hash_schemes,
                     ),
                 )
             )
@@ -140,7 +148,11 @@ class StarliteUsers:
                     forgot_path=self._config.password_reset_handler_config.forgot_path,
                     reset_path=self._config.password_reset_handler_config.reset_path,
                     service_dependency=get_service_dependency(
-                        self._config.user_model, self._config.role_model, self._config.user_service_class
+                        user_model=self._config.user_model,
+                        role_model=self._config.role_model,
+                        user_service_class=self._config.user_service_class,
+                        secret=self._config.secret,
+                        hash_schemes=self._config.hash_schemes,
                     ),
                 )
             )
@@ -151,7 +163,11 @@ class StarliteUsers:
                     user_create_dto=self._config.user_create_dto,
                     user_read_dto=self._config.user_read_dto,
                     service_dependency=get_service_dependency(
-                        self._config.user_model, self._config.role_model, self._config.user_service_class
+                        user_model=self._config.user_model,
+                        role_model=self._config.role_model,
+                        user_service_class=self._config.user_service_class,
+                        secret=self._config.secret,
+                        hash_schemes=self._config.hash_schemes,
                     ),
                 )
             )
@@ -168,7 +184,11 @@ class StarliteUsers:
                     role_update_dto=self._config.role_update_dto,  # type: ignore[arg-type]
                     user_read_dto=self._config.user_read_dto,
                     service_dependency=get_service_dependency(
-                        self._config.user_model, self._config.role_model, self._config.user_service_class
+                        user_model=self._config.user_model,
+                        role_model=self._config.role_model,
+                        user_service_class=self._config.user_service_class,
+                        secret=self._config.secret,
+                        hash_schemes=self._config.hash_schemes,
                     ),
                 )
             )
@@ -181,7 +201,11 @@ class StarliteUsers:
                     user_read_dto=self._config.user_read_dto,
                     user_update_dto=self._config.user_update_dto,
                     service_dependency=get_service_dependency(
-                        self._config.user_model, self._config.role_model, self._config.user_service_class
+                        user_model=self._config.user_model,
+                        role_model=self._config.role_model,
+                        user_service_class=self._config.user_service_class,
+                        secret=self._config.secret,
+                        hash_schemes=self._config.hash_schemes,
                     ),
                 )
             )
@@ -191,7 +215,11 @@ class StarliteUsers:
                     path=self._config.verification_handler_config.path,
                     user_read_dto=self._config.user_read_dto,
                     service_dependency=get_service_dependency(
-                        self._config.user_model, self._config.role_model, self._config.user_service_class
+                        user_model=self._config.user_model,
+                        role_model=self._config.role_model,
+                        user_service_class=self._config.user_service_class,
+                        secret=self._config.secret,
+                        hash_schemes=self._config.hash_schemes,
                     ),
                 )
             )