+++ tests/conftest.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timedelta
-from typing import TYPE_CHECKING, Generator
+from typing import TYPE_CHECKING, Any, Generator
 from unittest.mock import MagicMock
 from uuid import UUID, uuid4
 
@@ -52,7 +52,7 @@ password_manager = PasswordManager(hash_schemes=HASH_SCHEMES)
 
 
 class User(Base, SQLAlchemyUserMixin):  # type: ignore[valid-type, misc]
-    pass
+    __tablename__ = "user"
 
 
 class UserCreateDTO(BaseUserCreateDTO):
@@ -67,11 +67,8 @@ class UserUpdateDTO(BaseUserUpdateDTO):
     pass
 
 
-class UserService(BaseUserService[User, UserCreateDTO, UserUpdateDTO]):
-    user_model = User
-    role_model = None
-    secret = ENCODING_SECRET
-    hash_schemes = HASH_SCHEMES
+class UserService(BaseUserService[User, UserCreateDTO, UserUpdateDTO, Any]):
+    pass
 
 
 @pytest.fixture()
@@ -79,7 +76,7 @@ def admin_user() -> User:
     return User(
         id=UUID("01676112-d644-4f93-ab32-562850e89549"),
         email="admin@example.com",
-        password_hash=password_manager.get_hash(SecretStr("iamsuperadmin")),
+        password_hash=password_manager.hash(SecretStr("iamsuperadmin")),
         is_active=True,
         is_verified=True,
     )
@@ -90,7 +87,7 @@ def generic_user() -> User:
     return User(
         id=UUID("555d9ddb-7033-4819-a983-e817237b88e5"),
         email="good@example.com",
-        password_hash=password_manager.get_hash(SecretStr("justauser")),
+        password_hash=password_manager.hash(SecretStr("justauser")),
         is_active=True,
         is_verified=True,
     )
@@ -111,7 +108,7 @@ def unverified_user() -> User:
     return User(
         id=UUID("68dec058-b752-42eb-8e55-b94a7b275f99"),
         email="unverified@example.com",
-        password_hash=password_manager.get_hash(SecretStr("notveryverified")),
+        password_hash=password_manager.hash(SecretStr("notveryverified")),
         is_active=True,
         is_verified=False,
     )