+++ tests/test_role_integration/conftest.py
@@ -1,15 +1,19 @@
+from typing import List
 from uuid import UUID
 
 import pytest
 from pydantic import SecretStr
+from sqlalchemy import Column, ForeignKey
+from sqlalchemy.orm import relationship
+from sqlalchemy.orm.attributes import Mapped  # type: ignore[attr-defined]
 from sqlalchemy.orm.decl_api import declarative_base
 from starlite.middleware.session.memory_backend import MemoryBackendConfig
 
 from starlite_users import StarliteUsersConfig
+from starlite_users.adapter.sqlalchemy.guid import GUID
 from starlite_users.adapter.sqlalchemy.mixins import (
     SQLAlchemyRoleMixin,
-    SQLAlchemyUserRoleMixin,
-    UserRoleAssociationMixin,
+    SQLAlchemyUserMixin,
 )
 from starlite_users.config import RoleManagementHandlerConfig
 from starlite_users.guards import roles_accepted, roles_required
@@ -19,30 +23,33 @@ from starlite_users.schema import (
     BaseRoleUpdateDTO,
     BaseUserRoleReadDTO,
 )
-from starlite_users.service import BaseUserRoleService
+from starlite_users.service import BaseUserService
 from tests.conftest import UserCreateDTO, UserUpdateDTO, _Base, password_manager
 from tests.constants import ENCODING_SECRET
-from tests.utils import MockSQLAlchemyUserRoleRepository
+from tests.utils import MockSQLAlchemyUserRepository
 
 Base = declarative_base(cls=_Base)
 
 
-class User(Base, SQLAlchemyUserRoleMixin):  # type: ignore[valid-type, misc]
-    pass
+class User(Base, SQLAlchemyUserMixin):  # type: ignore[valid-type, misc]
+    __tablename__ = "user"
+
+    roles: Mapped[List["Role"]] = relationship("Role", secondary="user_role", lazy="joined")
 
 
 class Role(Base, SQLAlchemyRoleMixin):  # type: ignore[valid-type, misc]
+    __tablename__ = "role"
+
+
+class UserRole(Base):  # type: ignore[valid-type, misc]
+    __tablename__ = "user_role"
+
+    user_id = Column(GUID(), ForeignKey("user.id"))
+    role_id = Column(GUID(), ForeignKey("role.id"))
+
+
+class UserService(BaseUserService[User, UserCreateDTO, UserUpdateDTO, Role]):
     pass
-
-
-class UserRole(Base, UserRoleAssociationMixin):  # type: ignore[valid-type, misc]
-    pass
-
-
-class UserService(BaseUserRoleService[User, UserCreateDTO, UserUpdateDTO, Role]):
-    user_model = User
-    role_model = Role
-    secret = ENCODING_SECRET
 
 
 @pytest.fixture()
@@ -68,7 +75,7 @@ def admin_user(admin_role: Role) -> User:
     return User(
         id=UUID("01676112-d644-4f93-ab32-562850e89549"),
         email="admin@example.com",
-        password_hash=password_manager.get_hash(SecretStr("iamsuperadmin")),
+        password_hash=password_manager.hash(SecretStr("iamsuperadmin")),
         is_active=True,
         is_verified=True,
         roles=[admin_role],
@@ -80,7 +87,7 @@ def generic_user() -> User:
     return User(
         id=UUID("555d9ddb-7033-4819-a983-e817237b88e5"),
         email="good@example.com",
-        password_hash=password_manager.get_hash(SecretStr("justauser")),
+        password_hash=password_manager.hash(SecretStr("justauser")),
         is_active=True,
         is_verified=True,
         roles=[],
@@ -124,7 +131,7 @@ def mock_user_repository(
     writer_role: Role,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    UserRepository = MockSQLAlchemyUserRoleRepository
+    UserRepository = MockSQLAlchemyUserRepository
     user_store = {
         str(admin_user.id): admin_user,
         str(generic_user.id): generic_user,
@@ -136,6 +143,6 @@ def mock_user_repository(
     }
     monkeypatch.setattr(UserRepository, "user_store", user_store)
     monkeypatch.setattr(UserRepository, "role_store", role_store)
-    monkeypatch.setattr("starlite_users.service.SQLAlchemyUserRoleRepository", UserRepository)
-    monkeypatch.setattr("starlite_users.user_handlers.SQLAlchemyUserRoleRepository", UserRepository)
-    monkeypatch.setattr("starlite_users.dependencies.SQLAlchemyUserRoleRepository", UserRepository)
+    monkeypatch.setattr("starlite_users.service.SQLAlchemyUserRepository", UserRepository)
+    monkeypatch.setattr("starlite_users.user_handlers.SQLAlchemyUserRepository", UserRepository)
+    monkeypatch.setattr("starlite_users.dependencies.SQLAlchemyUserRepository", UserRepository)