+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+[0.4.0]
+
+- unify `BaseUserService` and `BaseUserRoleService` et al.
+- add configuration options to `StarliteUsersConfig`.
+- add `hash_schemes` configuration to `PasswordManager`
+- remove configuration class variables from `BaseUserService`.
+- remove `UserRoleAssociationMixin`.
+- remove `SQLAlchemyUserRoleMixin`.
+- remove `__tablename__` declarations from SQLAlchemy mixins.
+
+[0.3.3]
+
+- fix session login response serialization issue.
+
+[0.3.2]
+
+- fix verification issue.
+
+[0.3.1]
+
+- replace broken SQLAlchemy forward refs.
+
 [0.3.0]
 
 - add `BaseUserRoleService`